lesson7: Суперкласс Path имплементирован
