+++ src/main/java/lesson7/DepthFirstPath.java
@@ -2,43 +2,11 @@ package lesson7;
 
 import java.util.LinkedList;
 
-public class DepthFirstPath {
-    private boolean[] marked;
-    private int[] edgeTo;
-    private int source;
+public class DepthFirstPath extends Path {
 
     public DepthFirstPath(Graph g, int source) {
-        this.source = source;
-        edgeTo = new int[g.getVertexCount()];
-        marked = new boolean[g.getVertexCount()];
-        dfs(g, source);
+        super(g, source, "depth");
     }
 
-    private void dfs(Graph g, int v) {
-        marked[v] = true;
-        for (int w : g.getAdjList(v)) {
-            if (!marked[w]) {
-                edgeTo[w] = v;
-                dfs(g, w);
-            }
-        }
-    }
-
-    public boolean hasPathTo(int dist) {
-        return marked[dist];
-    }
-
-    public LinkedList<Integer> pathTo(int dist) {
-        if (!hasPathTo(dist)) {
-            return null;
-        }
-        LinkedList<Integer> stack = new LinkedList<>();
-        int vertex = dist;
-        while (vertex != source) {
-            stack.push(vertex);
-            vertex = edgeTo[vertex];
-        }
-        return stack;
-    }
 
 }
