+++ src/main/java/lesson7/Main.java
@@ -16,27 +16,27 @@ public class Main {
         graph.addEdge(6, 2);
         graph.addEdge(4, 2);
 
-        BreadthFirstPath bfp = new BreadthFirstPath(graph, 3);
+//        BreadthFirstPath bfp = new BreadthFirstPath(graph, 3);
 //        System.out.println(bfp.hasPathTo(1));
-        System.out.println(bfp.pathTo(1));
-        System.out.println(bfp.pathTo(4));
-        System.out.println(bfp.pathTo(7));
+//        System.out.println(bfp.pathTo(1));
+//        System.out.println(bfp.pathTo(4));
+//        System.out.println(bfp.pathTo(7));
+
+//
+
+ //       System.out.println(graph.getVertexCount());
+ //       System.out.println(graph.getEdgeCount());
+ //       System.out.println(graph.getAdjList(4));
+
+        DepthFirstPath dfp = new DepthFirstPath(graph, 2);
+        System.out.println(dfp.hasPathTo(0));
+        System.out.println(dfp.pathTo(0));
 
         BreadthFirstPath bfp1 = new BreadthFirstPath(graph, 5);
-//        System.out.println(bfp.hasPathTo(1));
+        System.out.println(bfp1.hasPathTo(1));
         System.out.println(bfp1.pathTo(1));
         System.out.println(bfp1.pathTo(4));
         System.out.println(bfp1.pathTo(3));
-
-
-
-//        System.out.println(graph.getVertexCount());
-//        System.out.println(graph.getEdgeCount());
-//        System.out.println(graph.getAdjList(4));
-
-//        DepthFirstPath dfp = new DepthFirstPath(graph, 2);
-//        System.out.println(dfp.hasPathTo(0));
-//        System.out.println(dfp.pathTo(0));
 
 
     }
