+++ src/main/java/lesson7/Path.java
@@ -1,0 +1,75 @@
+package lesson7;
+
+import java.util.LinkedList;
+
+class Path {
+    protected boolean[] marked;
+    protected int[] edgeTo;
+    protected int source;
+    protected String type = "";
+
+    public Path(Graph g, int source, String type) {
+        this.source = source;
+        edgeTo = new int[g.getVertexCount()];
+        marked = new boolean[g.getVertexCount()];
+        if (type == "depth") {
+            dfs(g, source);
+        } else {
+            bfs(g, source);
+        }
+    }
+
+    public void realise(Graph g, int v) {
+        marked[v] = true;
+        for (int w : g.getAdjList(v)) {
+            if (!marked[w]) {
+                edgeTo[w] = v;
+                realise(g, w);
+            }
+        }
+    }
+
+    private void bfs(Graph g, int source) {
+        LinkedList<Integer> queue = new LinkedList<>();
+        queue.addLast(source);
+        marked[source] = true;
+
+        while (!queue.isEmpty()) {
+            int vertex = queue.removeFirst();
+            for (int w : g.getAdjList(vertex)) {
+                if (!marked[w]) {
+                    marked[w] = true;
+                    edgeTo[w] = vertex;
+                    queue.addLast(w);
+                }
+            }
+        }
+    }
+
+    private void dfs(Graph g, int v) {
+        marked[v] = true;
+        for (int w : g.getAdjList(v)) {
+            if (!marked[w]) {
+                edgeTo[w] = v;
+                dfs(g, w);
+            }
+        }
+    }
+
+    public boolean hasPathTo(int dist) {
+        return marked[dist];
+    }
+
+    public LinkedList<Integer> pathTo(int dist) {
+        if (!hasPathTo(dist)) {
+            return null;
+        }
+        LinkedList<Integer> stack = new LinkedList<>();
+        int vertex = dist;
+        while (vertex != source) {
+            stack.push(vertex);
+            vertex = edgeTo[vertex];
+        }
+        return stack;
+    }
+}
